server/requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
djangorestframework
markdown
django-filter
requests